# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface User {
  id: string;
  full_name: string;
  email: string;
}

export interface SignUpInput {
  /**
     * @minLength 2
     * @maxLength 60
     */
  full_name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}

export interface SignInInput {
  email: string;
  password: string;
}

export type MessageRole = typeof MessageRole[keyof typeof MessageRole];


export const MessageRole = {
  user: 'user',
  assistant: 'assistant',
} as const;

export interface Message {
  id: string;
  role: MessageRole;
  content: string;
  created_at: string;
}

export interface ChatMessageInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}

export interface ChatReply {
  user_message: Message;
  assistant_message: Message;
}

export interface CvWorkExperience {
  company: string;
  role: string;
  start_date: string;
  end_date: string;
  responsibilities: string[];
}

export interface CvEducation {
  institution: string;
  degree: string;
  field: string;
  year: string;
}

export interface CvData {
  full_name: string;
  email: string;
  phone: string;
  location: string;
  linkedin: string;
  career_objective: string;
  work_experience: CvWorkExperience[];
  education: CvEducation[];
  skills: string[];
  certifications: string[];
  languages: string[];
  interests: string[];
}

export interface Cv {
  id: string;
  version: number;
  cv_data: CvData;
  created_at: string;
}

export interface CvGenerationInput {
  messages: Message[];
}