# lib/api-zod/src/generated/types/chatMessageInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ChatMessageInput {
  /**
     * @minLength 1
     * @maxLength 1000
     */
  content: string;
}