# lib/api-zod/src/generated/types/chatReply.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';

export interface ChatReply {
  user_message: Message;
  assistant_message: Message;
}