# lib/api-zod/src/generated/types/cvData.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CvEducation } from './cvEducation';
import type { CvWorkExperience } from './cvWorkExperience';

export interface CvData {
  full_name: string;
  email: string;
  phone: string;
  location: string;
  linkedin: string;
  career_objective: string;
  work_experience: CvWorkExperience[];
  education: CvEducation[];
  skills: string[];
  certifications: string[];
  languages: string[];
  interests: string[];
}