# lib/api-zod/src/generated/types/cvEducation.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CvEducation {
  institution: string;
  degree: string;
  field: string;
  year: string;
}