# lib/api-zod/src/generated/types/cvGenerationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Message } from './message';

export interface CvGenerationInput {
  messages: Message[];
}