# lib/api-zod/src/generated/types/cvWorkExperience.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CvWorkExperience {
  company: string;
  role: string;
  start_date: string;
  end_date: string;
  responsibilities: string[];
}