# lib/api-zod/src/generated/types/signUpInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SignUpInput {
  /**
     * @minLength 2
     * @maxLength 60
     */
  full_name: string;
  email: string;
  /** @minLength 8 */
  password: string;
}